% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{lambda}
\alias{lambda}
\title{Average recruitment curve}
\usage{
lambda(p1, p2, t, c2, c1, type)
}
\arguments{
\item{p1}{Shape parameter.}

\item{p2}{Rate parameter.}

\item{t}{Grid of time points.}

\item{c2}{Multiplicative constant.}

\item{c1}{Shifting constant.}

\item{type}{Either the pdf of a Gamma random variable (default) or the cdf.}
}
\value{
A vector with the average recruitment intensity at each time point.
}
\description{
This function generates the average recruitment curve according to either the pdf or cdf of a
Gamma random variable to be used in the \code{\link{sim_rec}} function to simulate recruitment data.
}
