% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{sim_rec}
\alias{sim_rec}
\title{Data generating function}
\usage{
sim_rec(
  alpha,
  p1,
  p2,
  c2,
  c1,
  type,
  t_p,
  Tmax,
  s_time,
  C_act,
  random_init = TRUE
)
}
\arguments{
\item{alpha}{Shape parameter of the Gamma distribution.}

\item{p1}{Shape parameter of the average recruitment intensity.}

\item{p2}{Rate parameter of the average recruitment intensity.}

\item{c2}{Multiplicative constant of the average recruitment intensity.}

\item{c1}{Shifting constant of the average recruitment intensity.}

\item{type}{Form of the average recruitment intensity. It can be either the pdf of a Gamma random variable (default) or the cdf.}

\item{t_p}{Plateau point.}

\item{Tmax}{Maximum value for the grid of time points.}

\item{s_time}{Vector of the centers' initiation times. Must be of length \code{C_act} if \code{random_init = F}.}

\item{C_act}{Number of active centers.}

\item{random_init}{Logical; if TRUE (default), the centers' initiation times are selected randomly from the values in \code{s_time}.}
}
\value{
A list with two elements: the vector of initiation times and a list where each element contains the recruitment data of one center
formatted as a data frame with two columns; the first column is the grid of recruitment days for the concerned center,
and the second column indicates the respective number of enrollments on that day.
}
\description{
Data generating function
}
\examples{
\dontrun{sim_data <- sim_rec(alpha = 1, p1 = 10, p2 = 0.15, c2 = 0.3, c1 = 0.1,
type = "cdf", t_p = 150, Tmax = 200, s_time = 1:100, C_act = 30, random_init = TRUE)}
}
