% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{tPG_plot}
\alias{tPG_plot}
\title{Display forecasted enrollments}
\usage{
tPG_plot(pred, xlim, ylim, print_T = F, main = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{pred}{Output from the \code{tPG_pred} function.}

\item{xlim}{Standard \code{plot} argument}

\item{ylim}{Standard \code{plot} argument}

\item{print_T}{Logical. If TRUE, point estimate and CrIs for the time to achieve the sample size are printed.}

\item{main}{Standard \code{plot} argument}

\item{xlab}{Standard \code{plot} argument}

\item{ylab}{Standard \code{plot} argument}
}
\value{
A plot of the recruitment process.
}
\description{
Display forecasted enrollments
}
\examples{
sim_data <- sim_rec(alpha = 1, p1 = 10, p2 = 0.15, c2 = 0.3, c1 = 0.1,
type = "cdf", t_p = 150, Tmax = 200, s_time = 1:100, C_act = 30, random_init = TRUE)

est <- tPG_est(data = sim_data$data, t_int = 200 , tp_start = c(20, 60, 100, 140), degree = c(2, 3), den = c(NA, 2, 3) )
pred <- tPG_pred(est = est, start_init = sim_data$start, N=5000, tpred = 201:900, conf_level = 0.95)
tPG_plot(pred = pred, xlim=c(0,550), ylim=c(0,5500), print_T = TRUE, xlab = "Days", ylab = "Enrollments", main = "Forecasted recruitments")
}
