% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{tPG_pred}
\alias{tPG_pred}
\title{Recruitment predictions via the tPG model}
\usage{
tPG_pred(est, tpred, start_init, N = NULL, start_add = NULL, conf_level = 0.95)
}
\arguments{
\item{est}{Output from the \code{tPG_est} function}

\item{tpred}{Future time point for predictions. It can be a single value or a grid of time points.}

\item{start_init}{Starting times of the centers already activated.}

\item{N}{Optional; target sample size. If specified, \code{tpred} must be a grid of values.}

\item{start_add}{Starting times of centers not yet activated.}

\item{conf_level}{Confidence level of the credible interval.}
}
\value{
This function returns multiple objects, including the (additional) predicted enrollments and associated credible intervals from the current time to the values
specified in \code{tpred}. If \code{N} is specified, the function also returns the estimated time left to recruit the target sample size.
}
\description{
Recruitment predictions via the tPG model
}
\examples{
\dontrun{sim_data <- sim_rec(alpha = 1, p1 = 10, p2 = 0.15, c2 = 0.3, c1 = 0.1,
type = "cdf", t_p = 150, Tmax = 200, s_time = 1:100, C_act = 30, random_init = TRUE)

est <- tPG_est(data = sim_data$data, t_int = 200 , tp_start = c(20, 60, 100, 140), degree = c(2, 3), den = c(NA, 2, 3) )
pred_1 <- tPG_pred(est = est, start_init = sim_data$start, N=5000, tpred = 201:600, conf_level = 0.95)}
}
