% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{tPG_pred_cov}
\alias{tPG_pred_cov}
\title{Recruitment predictions via the tPG model with covariates}
\usage{
tPG_pred_cov(
  est,
  X = X,
  tpred,
  start_init,
  N = NULL,
  start_add = NULL,
  conf_level = 0.95
)
}
\arguments{
\item{est}{Output from the \code{tPG_est} function}

\item{X}{Matrix of center-specific covariates where each row is a covariate.
The centers represented by row must be in the same chronologial order used in \code{data}.}

\item{tpred}{Future time point for predictions. It can be a single value or a grid of time points.}

\item{start_init}{Starting times of the centers already activated.}

\item{N}{Optional; target sample size. If specified, \code{tpred} must be a grid of values.}

\item{start_add}{Starting times of centers not yet activated.}

\item{conf_level}{Confidence level of the credible interval.}
}
\value{
This function returns multiple objects, including the (additional) predicted enrollments and associated credible intervals from the current time to the values
specified in \code{tpred}. If \code{N} is specified, the function also returns the estimated time left to recruit the target sample size.
}
\description{
Recruitment predictions via the tPG model with covariates
}
